Run matrix profile to the left and to the right
A new flag has been added to the OperationInfo which allows to reduce
the indexes to the left and to the right when running self_join. If the
flag is set to Reduce::LEFT_RIGHT the matrix profile to the left will be
returned in profile_a and the matrix profile to the right in profile_b.

The Profile structure which had a std::vector of ProfileData has been
modified and now it has a ProfileData member instead of std::vector
since it was being used always just the [0] position.